pyproject.toml
[project]
name = "make-ipinyou-data"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "devcomfort", email = "im@devcomfort.me" }
]
dependencies = [
    "python-dotenv>=1.2.1",
    "kagglehub>=0.3.13",
    "tqdm>=4.67.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/make_ipinyou_data"]
